enterprise-blockchain/public-files/package/requirements.txt
fastapi==0.79.0
uvicorn==0.17.6
requests==2.28.1
pyyaml==6.0
aiohttp[speedups]==3.8.3
web3==6.6.1
